requirements.txt
Babel==2.5.3
click==6.7
dnspython==1.15.0
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-PyMongo==0.5.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pymongo==3.6.1
python-dotenv==0.8.2
pytz==2018.4
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
